splitted quiz.jsx into components
