+++ src/components/QuizResults/QuizResults.css
@@ -1,0 +1,5 @@
+.result-container {
+   max-width: 600px;
+   padding: 2rem;
+   text-align: center;
+ }
